backend/requirements.txt
# Requirements for the Power Monitoring Dashboard backend
Django==5.1.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
firebase-admin==6.6.0